cmd submit and resource binding refactor

+++ FireAndForget.v2/FireAndForget_v2Main.cpp
@@ -26,7 +26,7 @@ void FireAndForget_v2Main::CreateRenderers(const std::shared_ptr<DX::DeviceResou
 	// TODO: Replace this with your app's content initialization.
 	m_sceneRenderer = std::unique_ptr<Renderer>(new Renderer(deviceResources));
 	rendererWrapper_.Init(m_sceneRenderer.get());
-	scene_.Init(&rendererWrapper_, deviceResources->GetOutputSize().Width, deviceResources->GetOutputSize().Height);
+	scene_.Init(&rendererWrapper_, (int)deviceResources->GetOutputSize().Width, (int)deviceResources->GetOutputSize().Height);
 	OnWindowSizeChanged();
 }
